test(apps): follow account test file conventions
